Refactored game package to be slightly easier to find symbols
