+++ world.json
@@ -159,7 +159,8 @@
 					"LIVING_ROOM",
 					"KITCHEN",
 					"LIVING_ROOM",
-					"FRONT_YARD"
+					"FRONT_YARD",
+					"BACKYARD"
 				]
 			},
 			"dialog": ["Grrrr :3"]
